refactor(git-utils): rename interface file name
