+++ packages/amplication-git-utils/src/git/git-factory.ts
@@ -1,4 +1,4 @@
-import { GitProvider } from "../GitProvider";
+import { GitProvider } from "../git-provider.interface.ts";
 import { EnumGitProvider, GitProviderArgs } from "../types";
 import { INVALID_SOURCE_CONTROL_ERROR_MESSAGE } from "./git.constants";
 import { GithubService } from "./github.service";
